% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{decode_colour}
\alias{decode_colour}
\title{Decode RGB hex-strings into colour values}
\usage{
decode_colour(colour, alpha = FALSE, to = "rgb", white = "D65")
}
\arguments{
\item{colour}{A character vector of hex-encoded values or a valid colour name
as given in \code{\link[grDevices:colours]{grDevices::colours()}}.}

\item{alpha}{If \code{TRUE} the alpha channel will be returned as well (scaled
between 0 and 1). If no alpha channel exists in the colour it will be
assumed 1. If \code{FALSE} any alpha channel is ignored.}

\item{to}{The output colour space. Allowed values are: \code{"cmy"},
\code{"cmyk"}, \code{"hsl"}, \code{"hsb"}, \code{"hsv"}, \code{"lab"} (CIE L*ab), \code{"hunterlab"}
(Hunter Lab), \code{"lch"} (CIE Lch(ab) / polarLAB), \code{"luv"}, \code{"rgb"} (sRGB),
\code{"xyz"}, \code{"yxy"}, or \code{"hcl"} (CIE Lch(uv) / polarLuv)}

\item{white}{The white reference of the output colour space. Will only have
an effect for relative colour spaces such as Lab and luv. Any value accepted
by \code{\link[=as_white_ref]{as_white_ref()}} allowed.}
}
\value{
A numeric matrix with a row for each element in \code{colour} and either
3, 4, or 5 columns depending on the value of \code{alpha} and \code{to}.
}
\description{
This is a version of \code{\link[grDevices:col2rgb]{grDevices::col2rgb()}} that returns the colour values in
the standard form expected by farver (matrix with a row per colour). As with
\code{\link[=encode_colour]{encode_colour()}} it can do colour conversion on the fly, meaning that you can
decode a hex string directly into any of the supported colour spaces.
}
\seealso{
Other encoding and decoding functions: \code{\link{encode_colour}}
}
\concept{encoding and decoding functions}
