% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farver-package.R
\docType{package}
\name{farver-package}
\alias{farver}
\alias{farver-package}
\title{farver: Vectorised Colour Conversion and Comparison}
\description{
One paragraph description of what the package does as one or more full sentences.
}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomasp85@gmail.com}

Authors:
\itemize{
  \item Berendea Nicolae (Author of the ColorSpace C++ library)
}

}
