% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/native.R
\name{native-encoding}
\alias{native-encoding}
\alias{encode_native}
\alias{decode_native}
\title{Convert to and from the R native colour representation}
\usage{
encode_native(colour, alpha = NULL, from = "rgb", white = "D65", na_value = NA)

decode_native(colour)
}
\arguments{
\item{colour}{For \code{encode_native} either a vector of hex-encoded
colours/colour names or a matrix encoding colours in any of the supported
colour spaces. For \code{decode_native} it is a vector of
integers.}

\item{alpha}{A numeric vector between 0 and 1. Will be recycled to the number
of rows in \code{colour}. If \code{NULL} or a single \code{NA} it will be ignored.}

\item{from}{The input colour space. Allowed values are: \code{"cmy"},
\code{"cmyk"}, \code{"hsl"}, \code{"hsb"}, \code{"hsv"}, \code{"lab"} (CIE L*ab), \code{"hunterlab"}
(Hunter Lab), \code{"oklab"}, \code{"lch"} (CIE Lch(ab) / polarLAB), \code{"luv"},
\code{"rgb"} (sRGB), \code{"xyz"}, \code{"yxy"} (CIE xyY), \code{"hcl"} (CIE Lch(uv) / polarLuv),
or \code{"oklch"} (Polar form of oklab)}

\item{white}{The white reference of the input colour space. Will only have an
effect for relative colour spaces such as Lab and luv. Any value accepted by
\code{\link[=as_white_ref]{as_white_ref()}} allowed.}

\item{na_value}{A valid colour string or \code{NA} to use when \code{colour} contains
\code{NA} elements or is invalid in the RGB space. The general approach in farver
is to carry \code{NA} values over, but if you want to mimick \code{\link[=col2rgb]{col2rgb()}} you should
set \code{na_value = 'transparent'}, i.e. treat \code{NA} as transparent white.}
}
\value{
\code{encode_native()} returns an integer vector and \code{decode_native()}
returns a character vector, both matching the length of the input.
}
\description{
Colours in R are internally encoded as integers when they are passed around
to graphics devices. The encoding splits the 32 bit in the integer between
red, green, blue, and alpha, so that each get 8 bit, equivalent to 256
values. It is very seldom that an R user is subjected to this representation,
but it is present in the \code{nativeRaster} format which can be obtained from
e.g. capturing the content of a graphic device (using \code{dev.cap()}) or reading
in PNG files using \code{png::readPNG(native = TRUE)}. It is very rare that you
might need to convert back and forth between this format, but it is provided
here for completeness.
}
\examples{

# Get native representation of navyblue and #228B22
native_col <- encode_native(c('navyblue', '#228B22'))
native_col

# Convert back
decode_native(native_col)

}
