% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/native.R
\name{replace_alpha_native}
\alias{replace_alpha_native}
\title{Replace the alpha channel of a native vector}
\usage{
replace_alpha_native(colour, alpha)
}
\arguments{
\item{colour}{An integer vector, expressing colours in native format}

\item{alpha}{A numeric vector between 0 and 1. Will be recycled to the number
of rows in \code{colour}. If \code{NULL} or a single \code{NA} it will be ignored.}
}
\value{
The \code{colour} integer vector, with updated alpha values
}
\description{
Replace the alpha channel of a native vector
}
\examples{
# Get native representation of navyblue and #228B22
native_col <- encode_native(c('navyblue', '#228B22'))
# Replace the alpha of the native colour to 0.5:
native_with_alpha <- replace_alpha_native(native_col, 0.5)

}
