% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farver-package.R
\docType{package}
\name{farver-package}
\alias{farver}
\alias{farver-package}
\title{farver: Vectorised Colour Conversion and Comparison}
\description{
The encoding of colour can be handled in many different ways, using
different colour spaces. As different colour spaces have different uses,
efficient conversion between these representations are important. The
'farver' package provides a set of functions that gives access to very fast
colour space conversion and comparisons implemented in C++, and offers
100-fold speed improvements over the 'convertColor' function in the
'grDevices' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thomasp85/farver}
  \item Report bugs at \url{https://github.com/thomasp85/farver/issues}
}

}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomasp85@gmail.com}

Authors:
\itemize{
  \item Berendea Nicolae (Author of the ColorSpace C++ library)
}

}
