% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farver-package.R
\docType{package}
\name{farver-package}
\alias{farver}
\alias{farver-package}
\title{farver: High Performance Colour Space Manipulation}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

The encoding of colour can be handled in many different ways, using
    different colour spaces. As different colour spaces have different uses,
    efficient conversion between these representations are important. The 
    'farver' package provides a set of functions that gives access to very fast
    colour space conversion and comparisons implemented in C++, and offers 
    speed improvements over the 'convertColor' function in the 'grDevices' 
    package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://farver.data-imaginist.com}
  \item \url{https://github.com/thomasp85/farver}
  \item Report bugs at \url{https://github.com/thomasp85/farver/issues}
}

}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomasp85@gmail.com} (\href{https://orcid.org/0000-0002-5147-4711}{ORCID})

Authors:
\itemize{
  \item Berendea Nicolae (Author of the ColorSpace C++ library)
  \item Romain François \email{romain@purrple.cat} (\href{https://orcid.org/0000-0002-2444-4226}{ORCID})
}

}
\keyword{internal}
